+++ ui/src/app/components/buttons.tsx
@@ -680,16 +680,3 @@ export class SlidingFabReact extends React.Component<
     );
   }
 }
-
-interface CloseButtonProps {
-  onClose: Function;
-  style?: CSSProperties;
-}
-export const CloseButton = (props: CloseButtonProps) => {
-  const { onClose, style } = props;
-  return (
-    <Clickable {...{ style }} onClick={onClose}>
-      <img src={times} style={{ height: '27px', width: '17px' }} alt='Close' />
-    </Clickable>
-  );
-};
